Moved function save_dataset_description to exportutils module
